services/marketplace-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
pymysql==1.1.1
pydantic==2.9.0
strawberry-graphql[fastapi]==0.243.0
